Created base class for logic
