+++ SpiceBot/BasedCringe/Opinion.cs
