+++ SpiceBot/NumberStationLogic.cs
@@ -1,0 +1,6 @@
+namespace SpiceBot
+{
+    internal class NumberStationLogic : SpiceLogic
+    {
+    }
+}
